Add the button & move file to folder academic-industry components 🐯
